refactor(core): add `StyleRoot` concept
This represents a unique location where styles can be rendered and will apply to distinct elements. This is defined as simply `Document | ShadowRoot`. Styles rendered to a `Document` should go in its `HTMLHeadElement` while styles applied to a `ShadowRoot` should just be rendered within that root.

This needs to be public API as it will be added to `Renderer2` which is already public API.