wip: add `RNode.prototype.getRootNode`
This exposes [`Node.prototype.getRootNode`](https://developer.mozilla.org/en-US/docs/Web/API/Node/getRootNode) to the framework as a mechanism of identifying the containing `ShadowRoot` of an element.

The practical type is `HTMLDocument | ShadowRoot`, but TS DOM types classify this as `Node`, so we need to align with that.